refactor to vim.pack take 1
