+++ src/nvim/plugin/zz-finalize.lua
@@ -1,0 +1,3 @@
+local vimrc = require('vimrc')
+
+vimrc.run_plugin_setups()
